pyproject.toml
[tool.poetry]
name = "football-project"
version = "0.1.0"
description = ""
authors = ["AndreFelippeVidal <andre_vidal@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
ratelimit = "^2.2.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
